pyproject.toml
[project]
name = "content-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.120.1",
    "google-search-results>=2.4.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "nest-asyncio>=1.6.0",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "serpapi>=0.1.5",
    "streamlit>=1.45.1",
]
